+++ packages/orq-rc/RELEASES.md
@@ -4479,3 +4479,13 @@ Based on:
 - [typescript v3.14.0-rc.48] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.48 - packages/orq-rc
+
+## 2025-10-23 06:55:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.638.1 (2.729.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.49] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.49 - packages/orq-rc